[rename]user -> balance로 변경
