+++ src/main/java/kr/hhplus/be/server/presentation/balance/BalanceRequest.java
@@ -1,6 +1,6 @@
-package kr.hhplus.be.server.presentation.user;
+package kr.hhplus.be.server.presentation.balance;
 
-public class UserRequest {
+public class BalanceRequest {
     public record Charge(
         long user_id,
         long charge_amount
